deleted FavouriteExercise and replaced it by Exercise

+++ src/main/java/pl/workout/kowalczyk/com/app/model/data/dao/BaseDao.java
@@ -15,5 +15,5 @@ public interface BaseDao<T> {
 
     void delete(T domain);
 
-    T get(Long id);
+    T get(int id);
 }
